requirements.txt
fastapi
uvicorn
python-nmap
groq
requests
python-dotenv
aiohttp
pywin32
networkx